+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Master](https://github.com/appwise-labs/AppwiseCore)
 
+### New Features
+
+* Core: split off the application delegate and services into a separate subspec `UIApplication`, so that the `Core` subspec can be used in application extensions.
+
 ## [0.10.6](https://github.com/appwise-labs/AppwiseCore/releases/tag/0.10.6)
 
 ### Bug Fixes